adjusting the new version to work with ngrok and communicate with telegram

+++ src/main/java/com/ufps/cryptobot/contract/NewsMessage.java
@@ -8,10 +8,11 @@ import lombok.Setter;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter @Setter
+@Getter
+@Setter
 public class NewsMessage {
     @JsonAlias("chat_id")
-    private Long chatID;
+    private Integer chatID;
     private String title;
     private String link;
 }
